force python 2.x for planet
ubuntu-latest has been upgraded to ubuntu-20.04, which removes a lot of
python2 support (including pip for python2)

On the other side, planet hasn't been updated since 2006,
and uses syntax incompatible with python 3.

for ex. `0666` which fails with:
```
SyntaxError: leading zeros in decimal integer literals are not permitted; use an 0o prefix for octal integers
```

So use ubuntu-18.04 - since it supports python2 better